requirements.txt
steam
discord.py
sqlalchemy
mysql-connector-python